refactor: toOtherSign 컴포넌트 생성, hover 이벤트 추가
