+++ src/pages/SignUpPage.tsx
@@ -1,11 +1,9 @@
-import { Box, Center, Text } from '@chakra-ui/react';
-import styled from '@emotion/styled';
-import { Link } from 'react-router-dom';
+import { Box } from '@chakra-ui/react';
 
 import LargeLogo from '@/components/base/LargeLogo';
 import BackNavigation from '@/components/navigation/BackNavigation';
-import SignUpForm from '@/components/signUp/SignUpForm';
-import ROUTES from '@/utils/constants/routes';
+import SignUpForm from '@/components/userSign/signUp/SignUpForm';
+import ToOtherSign from '@/components/userSign/ToOtherSign';
 
 const SignUpPage = () => {
   return (
@@ -13,21 +11,9 @@ const SignUpPage = () => {
       <BackNavigation />
       <LargeLogo src='/logo-lg.svg' />
       <SignUpForm />
-      <Center>
-        <Text mt='2rem' fontSize='md' fontWeight='medium' color='gray'>
-          이미 회원이신가요?
-          <StyledLink style={{ color: 'blue' }} to={ROUTES.SIGNIN}>
-            로그인
-          </StyledLink>
-        </Text>
-      </Center>
+      <ToOtherSign />
     </Box>
   );
 };
-
-const StyledLink = styled(Link)`
-  color: blue;
-  margin-left: 0.5rem;
-`;
 
 export default SignUpPage;
